feat: 完善编译工具链和自动化流程
- 新增 tools/setup_env.sh 工具链自动检测脚本（跨平台）
  - 自动搜索 ARM GCC / OpenOCD / Make / Python
  - 支持 PATH → 环境变量 → 常见路径 三级检测
  - 未安装时提供精确下载链接
  - 结果缓存到 tools/tools_paths.conf

- 新增 examples/dht11_oled_mqtt/build_flash_monitor.sh
  - 一键 compile / flash / monitor / all / clean
  - 自动读取 setup_env 配置，兜底搜索常见路径

- 更新 build.bat / build_and_flash.bat / Makefile
  - 移除硬编码路径，改为多级自动搜索
  - 添加 -Wno-missing-braces 消除字体数据警告

- 更新 tools/serial_monitor.py
  - 支持命令行指定端口和波特率
  - 新增 --list 列出可用串口
  - 改进错误提示

- 修复 main.c 中 sprintf 格式化警告 (%d → %lu)
- 修复 startup 文件缺少末尾换行符
- 更新 .gitignore 排除编译产物和工具链缓存
- 完善 README 文档：工具链检测、操作流程、故障排除

+++ examples/dht11_oled_mqtt/User/main.c
@@ -79,9 +79,9 @@ int main(void) {
             OLED_ShowNum(2, 9, data.humidity_dec, 1);
             OLED_ShowString(2, 10, "%");
             
-            sprintf(send_buffer, "{\"temperature\":%d.%d,\"humidity\":%d.%d,\"count\":%d}", 
-                    data.temperature, data.temperature_dec, 
-                    data.humidity, data.humidity_dec, counter);
+            sprintf(send_buffer, "{\"temperature\":%d.%d,\"humidity\":%d.%d,\"count\":%lu}",
+                    data.temperature, data.temperature_dec,
+                    data.humidity, data.humidity_dec, (unsigned long)counter);
             
             ret = ESP8266_ConnectTCP(SERVER_IP, SERVER_PORT);
             if (ret == 0) {
